部分bug微调，如删除提示，刷新列表，上传窗口样式等

+++ src/renderer/components/modules/property-manager.vue
@@ -50,10 +50,14 @@
 
       deleteBucket: function () {
         if (!this.currentBucket) return
-        this.$store.dispatch('bucket/deleteBucket', this.currentBucket).then(() => {
-          this.$emit('freshBucket')
-          this.$emit('closeManage')
-        })
+
+        this.$confirm('确定要删除吗？删除后数据不可恢复', '提示').then(() => {
+          this.$store.dispatch('bucket/deleteBucket', this.currentBucket).then(() => {
+            this.$emit('freshBucket')
+            this.closeDialog()
+          })
+        }).catch(() => {})
+
       }
     }
